+++ tensorcircuit/torchnn.py
@@ -87,8 +87,8 @@ class QuantumNet(torch.nn.Module):  # type: ignore
                 initf = torch.randn
             self.q_weights.append(torch.nn.Parameter(initf(ws)))  # type: ignore
 
-    def forward(self, inputs: Tensor) -> Tensor:
-        ypred = self.f(inputs, *self.q_weights)
+    def forward(self, *inputs: Tensor) -> Tensor:
+        ypred = self.f(*inputs, *self.q_weights)
         return ypred